Update Posh-Cisco.psm1
Fix for 'Get-CiscoSSHResponse' function and dependent functions to use the current version of Posh-SSH (v3.2.4). Move from 'Invoke-SSHCommand' to 'Invoke-SSHStreamShellCommand'.
Due to the move to Invoke-SSHStreamShellCommand, which can insert pauses (more... ) in the output, it provides an option to set 'terminal length 0'.
Set the default timeout to 5 seconds if output hangs, but the value can be overridden. This was added to Posh-SSH v3.2.2.